Create utils for pretty output when running gen-assets child processes locally

+++ polaris.shopify.com/scripts/gen-site-map.mjs
@@ -1,21 +1,41 @@
-import {execa} from 'execa';
-import path from 'path';
+import {prettyExeca, startLocalServer} from './util.mjs';
 
 const genSiteMap = async () => {
   const outputFile = 'public/sitemap.xml';
 
-  const nextBin = path.join(process.cwd(), 'node_modules/.bin/next');
-  const server = execa(nextBin, ['dev']);
+  const server = startLocalServer('dev');
+  server.catch((err) => {
+    if (!server.killed) {
+      console.error(err.stderr);
+      process.exit(-1);
+    }
+  });
 
-  const {stdout} = await execa('npx', [
-    'get-site-urls',
-    'http://localhost:3000',
-    `--output=${outputFile}`,
-    '--alias=https://polaris.shopify.com',
-  ]);
-  console.log(stdout);
-
-  await server.kill();
+  try {
+    await prettyExeca(
+      'npx',
+      [
+        'get-site-urls',
+        'http://localhost:3000',
+        `--output=${outputFile}`,
+        '--alias=https://polaris.shopify.com',
+      ],
+      {
+        stdout: 'inherit',
+        stderr: 'inherit',
+        pretty: {
+          text: 'Generating sitemap',
+          successText: 'Sitemap generated',
+          failText: 'Unable to generate sitemap',
+        },
+      },
+    );
+  } finally {
+    if (server) {
+      // Kill the yarn command _and_ the child process running node
+      await server.kill();
+    }
+  }
 };
 
 export default genSiteMap;
